requirements.txt
asgiref==3.3.4
Django==3.0.5
djangorestframework==3.12.4
djongo==1.3.4
Markdown==3.3.4
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4
